Tracker Patches 2835099 + 2835094 from freddie-chopin applied
currently bitOffset is somewhat "reversed". bitoffset="0" in xml config
file means the MSB, and that is displayed in the view as "bit 31". This
patch fixes that behaviour - bitoffset="0" means LSB and will be displayed
as "bit 0", whereas bitoffset="31" will be MSB, and will be displayed as
"bit 31". Another patch is going to follow, as one of xml files needs to be
fixed to conform to this scheme